Asserts: Handle asserts outside phases
Asserts outside phases are test bugs. So far we have emited a warning
message. Now, we create on-the-fly phase, we put the standalone
assert into it, and add an explicit failed assert with the message
about the presence of test bug (so these are actually propagated to
Beaker: previously, we emited only the message, which was not propagated
anywhere)