Reverse the library import order
Previously, the library import order preferred roots of the dependency
graph. These libraries can depend on those closer to the leaves,
and therefore it is better to start importing from the
dependency leaves as this helps to prevent a situation where
a library calls a function from a dependency library in its
verifier function.

It does not rule out this possibility completely, though.