Check for successful processing of library names
Previously, the hashtable replacement commands did not check
for success. In case bad syntax bash was supplied to the eval
commands, such as VAR-WITH-COMMA=CONTENT, causing it to fail, this
failure was not reflected in the import execution. This patch checks
for successful evaluation, and propagates the failure to the import
machinery.