rlRun: Robustify /dev/null usage
Previously, when a creation of a temporary file needed for some
operations failed, /dev/null was used instead. In these cases /dev/null
was propagated to places where it was passed to the user or even
deleted.

Now the failure is detected, reported and the features needing the files
are switched off for safety.