storage.sh: Internal persistent storage for BL
Often, different pieces of BL need to temporarily (but persistently)
store information. The example is rlService* function family, which
stores the original status of a service, so it can be restored at the
end of the test. Until now, different pieces were doing this
differently (dropping files, setting bash variables). This commit
introduces the internal storage API with namespace/key/value, and
migrates rlFileBackup and rlService to it.