rlCleanupGenFinal: provide non-env variables for the cleanup script
Commit fda79f594114c incorrectly assumed that all internal beakerlib
variables are exported to the environment. This caused the cleanup
execution to "work" on the basic level, but things using ie.
$__INTERNAL_JOURNALIST (and many others) were broken.

The solution is to extract even local/global "not exported" variables,
ideally with correct attributes (array, readonly, exported, ..).

Also fix documentation, which is incorrect even prior to this commit.

Signed-off-by: Jiri Jaburek <jjaburek@redhat.com>