journal rewrite
To reduce journal processing complexity from exponencial to linear we decided to
rewrite journal completely. Now we create internal meta file which is simply
appended and which is used to generate final journal in xml format at the ond of
the test run.

Regarding to this change we also changed / enhanced all the logs printed. Now we
put timestamp to all log messages and we print also the reverity all the time.
Thus logs are now consistent.

We've also decided to changle multiline log behaviour. Now we do not prefix all
the lines nor during the test neither in the final report.

There's also another regression which is planned to fix later. There's no effect
of setting LOG_LEVEL. As the journal is processed one-way we need to implement
it completely at bash side. We've decided not to block the release on this and
complement consistent mechanish later.