prevent repetitive sourcing
Previousely some files were sourced multiple times. This can be either
slow down sourcing of beakerlib or it can be destructive if some code
is executed while sourcing and does not count with that.

The recursive sourcing is there from historical reasons. Some parts of
beakerlib could be sourced separately to provide some functionality.