silence status of service in rlService{Start,Stop,Restore} functions (#49)
* silence status of service in rlService{Start,Stop,Restore} functions

The service status printed by rlServiceStart is actually the status of the
previous service run which is often misleading to the user.

I we want to print the status it should be relevant to the currently starting
service run, eg. status after the service is started.

* added a new function rlServiceStatus

Introduced this function as a compentation of removing the service status
from rlService{Start,Stop,Restore}.
Now, the user can print it or process it, including an exit code, whenever
needed.