do not reset dir attributes for backup dir itself
Previously, if a file was backed up the backup path was updated
with the real permission and other attributes for each patch component.
This was done even for "/" which causes the backup dir to lose write
persmission.
Now, the backup dir itself is not updated according to "/" which should
not cause any issues, anyway.