better code readability
Cleaned up the code from the COMPONENT/LIBRARY format.
Now, only LIBRARY is ised everywhere the library is referenced.
The last place where the LIBRARY is broken down to COMPONENT/LIBRARY
is the actual search for to library location where the path component
'Library' needs to be inserted into of the library name.