change the rpm download method order from 'direct yum' to 'yum direct'
If the direct method is not configured properly to check the right URL base it
unnecesasrily fails with the direct method while the yum method works right away.

Also, there might be an issue with having locally prepared yum repo holding
the same NVR as it is available on the Koji. The direct method would prefer
the Koji one. This is often the case for the scratch builds tested in
the CI systems.

Chaning the order will also speedup the process e.g. on CentOS Stream.